conf/bblayers.conf: reorg
I much prefer the idea of separating out the layers by category and then
adding each category into the BBLAYERS variable. However, doing it this way
makes it impossible to use the "bitbake-layers" command to remove layers. So
for the time-being I think it would be better to just add all the layers into
BBLAYERS, flat, so they can be manipulated.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>