setup: BUILDDIR improvements
Move the definition of BUILDDIR later until just before it is needed. This
avoids polluting the shell with this variable if the script ends up erroring
out.

Make BUILDDIR's definition an absolute path (this is required for devtool).

Append the DISTRO to the default BUILDDIR path (i.e. if the user doesn't
specify one themselves).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>