setup: add 'deactivate' script
Add a 'deactivate' function to the environment so when the user is done with
OE stuff they can return their environment to just about how it was before
they sourced the 'setup' script.

Some things will be different such as OLDPWD and possibly other things of
which the script isn't aware, but for the most part it works.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>